Revert "gh-10 game screen: offline mode"
This reverts commit e2986c73d534045c188a7e5010cf0022bcb5f25c.

Adding a cache manifest triggers the following error on Mobile Safari (iPhone/iPad) when SoundJS register the sound (only in an optimized build aka scripts bundled by concat/requirejs):
   SyntaxError: DOM Exception 12: An invalid or illegal string was specified.

On line 1 of createjs.Sound.handleLoad(), this.request.response passed to this.owner.context.decodeAudioData() is undefined. Moreover, this.request.responseText is [DOM Exception]. It is unclear what the invalid or illegal string is and in what context it appears.